Normalize database, persist OTP codes, and unify mobile UI

+++ backend/src/main/java/com/beni/dto/WalletResponse.java
@@ -1,16 +1,18 @@
 
 package com.beni.dto;
 
+import java.math.BigDecimal;
+
 public class WalletResponse {
 
     public Integer walletId;
     public Integer driverId;
-    public Double balance;
+    public BigDecimal balance;
 
     public WalletResponse() {
     }
 
-    public WalletResponse(Integer walletId, Integer driverId, Double balance) {
+    public WalletResponse(Integer walletId, Integer driverId, BigDecimal balance) {
         this.walletId = walletId;
         this.driverId = driverId;
         this.balance = balance;
